So, did a clean install of Windows on my system due to some weird errors I had been dealing with for too long
As a result this project has been upgraded to Unity 2018.x
Unity now does not natively support substance files which I had been using for the grass texture,
so that has now been updated to a new asset.

Aside from those base changes due to my reinstalling everything:

-Simplified ground collision check audio
-Added a timer for how long the ball has not been in contact with ground that is part of the hole
-Added a variable to store the velocity of the ball from the previous Update frame for more accurate use in collisions and audio relating to those collisions.

Next addition will be the mechanic to reset the ball to its previous position
in the event that it either (1) ends movement in an area not part of the current hole,
or (2) has not been in contact with part of the ground for that hole in x amount of time.