Zachary Mitchell - 2/20/2019
-Implemented UI with transition for when the ball is out of bounds

I'll be spending a bit of time on the basics of learning Blender for a few things I'd like to do for this project.  I don't plan to spend more than about a 2 weeks maximum learning Blender and making the basics of what I want to continue on this project.  My plan is to get the core of this game done in the near future, so I can at least have a playable version to show off.

All that said, I am moving in a few days, but hopefully, there aren't too many UNFORESEEN CONSEQUENCES that eat up all my free time.