Zach Mitchell - 5/5/2019
Updates:

-Implementation of basic working game loop with 2 holes

-Working instantiation and setup of player prefab

-Addition of GolfHole and GolfCourse scripts for game loop purposes

-Added vertical camera control

-Overall refining of scripts

_________________________________________

My goal is to have a proper and well documented game loop where the player can progress through the holes of a course by the end of tomorrow.

After, the loop itself is working, I can set up the UI and code for scoring.

My overall plan is to get the most simplistic version of the game done and working first.
After that, I can add a hiscore system, menu/options, multiplayer, etc.